Fix for lost content in sample app
onCreateView is not the correct place to restore the tab content because
it is not guaranteed to be called. For example, if the user opens the app,
swipes through all of the tabs, rotates the device twice, then swipes
though the tabs again, the content of some tabs will have been lost,
resulting in the user seeing "???". Moving the content restore to onCreate,
which is always called after a rotation, will prevent the content from
being lost.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>